refactor(i18n/scripts): factorize check-translation.js and translate-rules.js
